Switch from OC::$server->get to OCP\Server::get
And add a bit more typing to some classes + psalm issues

+++ lib/AppInfo/Application.php
@@ -36,6 +36,7 @@ use OCA\Deck\Db\AclMapper;
 use OCA\Deck\Db\AssignmentMapper;
 use OCA\Deck\Db\BoardMapper;
 use OCA\Deck\Db\CardMapper;
+use OCA\Deck\Db\User;
 use OCA\Deck\Event\AclCreatedEvent;
 use OCA\Deck\Event\AclDeletedEvent;
 use OCA\Deck\Event\AclUpdatedEvent;
@@ -62,15 +63,16 @@ use OCP\Comments\CommentsEntityEvent;
 use OCP\Comments\ICommentsManager;
 use OCP\EventDispatcher\Event;
 use OCP\EventDispatcher\IEventDispatcher;
+use OCP\Group\Events\GroupDeletedEvent;
 use OCP\IConfig;
 use OCP\IDBConnection;
-use OCP\IGroup;
 use OCP\IGroupManager;
-use OCP\IServerContainer;
-use OCP\IUser;
+use OCP\IRequest;
+use OCP\Server;
 use OCP\IUserManager;
 use OCP\Notification\IManager as NotificationManager;
 use OCP\Share\IManager;
+use OCP\User\Events\UserDeletedEvent;
 use OCP\Util;
 use Psr\Container\ContainerInterface;
 
@@ -79,13 +81,8 @@ class Application extends App implements IBootstrap {
 
 	public const COMMENT_ENTITY_TYPE = 'deckCard';
 
-	/** @var IServerContainer */
-	private $server;
-
 	public function __construct(array $urlParams = []) {
 		parent::__construct(self::APP_ID, $urlParams);
-
-		$this->server = \OC::$server;
 	}
 
 	public function boot(IBootContext $context): void {
@@ -141,33 +138,43 @@ class Application extends App implements IBootstrap {
 
 	private function registerUserGroupHooks(IUserManager $userManager, IGroupManager $groupManager): void {
 		$container = $this->getContainer();
+		/** @var IEventDispatcher $eventDispatcher */
+		$eventDispatcher = $container->get(IEventDispatcher::class);
 		// Delete user/group acl entries when they get deleted
-		$userManager->listen('\OC\User', 'postDelete', static function (IUser $user) use ($container) {
+		$eventDispatcher->addListener(UserDeletedEvent::class, static function (Event $event) use ($container): void {
+			if (!($event instanceof UserDeletedEvent)) {
+				return;
+			}
+			$user = $event->getUser();
 			// delete existing acl entries for deleted user
 			/** @var AclMapper $aclMapper */
-			$aclMapper = $container->query(AclMapper::class);
+			$aclMapper = $container->get(AclMapper::class);
 			$acls = $aclMapper->findByParticipant(Acl::PERMISSION_TYPE_USER, $user->getUID());
 			foreach ($acls as $acl) {
 				$aclMapper->delete($acl);
 			}
 			// delete existing user assignments
-			$assignmentMapper = $container->query(AssignmentMapper::class);
+			$assignmentMapper = $container->get(AssignmentMapper::class);
 			$assignments = $assignmentMapper->findByParticipant($user->getUID());
 			foreach ($assignments as $assignment) {
 				$assignmentMapper->delete($assignment);
 			}
 
 			/** @var BoardMapper $boardMapper */
-			$boardMapper = $container->query(BoardMapper::class);
+			$boardMapper = $container->get(BoardMapper::class);
 			$boards = $boardMapper->findAllByOwner($user->getUID());
 			foreach ($boards as $board) {
 				$boardMapper->delete($board);
 			}
 		});
 
-		$groupManager->listen('\OC\Group', 'postDelete', static function (IGroup $group) use ($container) {
+		$eventDispatcher->addListener(GroupDeletedEvent::class, static function (Event $event) use ($container): void {
+			if (!($event instanceof GroupDeletedEvent)) {
+				return;
+			}
+			$group = $event->getGroup();
 			/** @var AclMapper $aclMapper */
-			$aclMapper = $container->query(AclMapper::class);
+			$aclMapper = $container->get(AclMapper::class);
 			$aclMapper->findByParticipant(Acl::PERMISSION_TYPE_GROUP, $group->getGID());
 			$acls = $aclMapper->findByParticipant(Acl::PERMISSION_TYPE_GROUP, $group->getGID());
 			foreach ($acls as $acl) {
@@ -181,6 +188,7 @@ class Application extends App implements IBootstrap {
 			$event->addEntityCollection(self::COMMENT_ENTITY_TYPE, function ($name) {
 				/** @var CardMapper */
 				$cardMapper = $this->getContainer()->get(CardMapper::class);
+				/** @var PermissionService $permissionService */
 				$permissionService = $this->getContainer()->get(PermissionService::class);
 
 				try {
@@ -203,7 +211,7 @@ class Application extends App implements IBootstrap {
 		$resourceManager->registerResourceProvider(ResourceProviderCard::class);
 
 		$symfonyAdapter->addListener('\OCP\Collaboration\Resources::loadAdditionalScripts', static function () {
-			if (strpos(\OC::$server->getRequest()->getPathInfo(), '/call/') === 0) {
+			if (strpos(Server::get(IRequest::class)->getPathInfo(), '/call/') === 0) {
 				// Talk integration has its own entrypoint which already includes collections handling
 				return;
 			}
